development/wol-user-service/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
sqlalchemy==1.4.46
pydantic==1.10.8
psycopg2-binary==2.9.6  # For PostgreSQL support
python-dotenv==1.0.0 # Optional, for managing environment variables
pytest==8.3.4
python-jose==3.3.0
passlib==1.7.4